move git repository db lib to grlib
