+++ grlib/db/gitrepository/v1/client_delete_git_repository_by_id.go
@@ -1,7 +1,7 @@
 package grcgitrepositorydb
 
 import (
-	grcgitrepositorydbdaos "github.com/kazekim/backend-engineer-challenge/grchallengeapi/pkgs/db/gitrepository/v1/daos"
+	grcgitrepositorydbdaos "github.com/kazekim/backend-engineer-challenge/grlib/db/gitrepository/v1/daos"
 	"github.com/kazekim/backend-engineer-challenge/grlib/grerrors"
 )
