+++ grlib/db/gitrepository/v1/client_list_git_repository_scan_results.go
@@ -1,8 +1,8 @@
 package grcgitrepositorydb
 
 import (
-	grcgitrepositorydbdaos "github.com/kazekim/backend-engineer-challenge/grchallengeapi/pkgs/db/gitrepository/v1/daos"
 	"github.com/kazekim/backend-engineer-challenge/grlib/besqlx"
+	grcgitrepositorydbdaos "github.com/kazekim/backend-engineer-challenge/grlib/db/gitrepository/v1/daos"
 	"github.com/kazekim/backend-engineer-challenge/grlib/grerrors"
 )
