+++ grlib/db/gitrepository/v1/daos/git_repository_scan_result_filter.go
@@ -1,11 +1,11 @@
 package grcgitrepositorydbdaos
 
 import (
-	grcenums "github.com/kazekim/backend-engineer-challenge/grchallengeapi/enums"
+	"github.com/kazekim/backend-engineer-challenge/grlib/grenums"
 )
 
 type GitRepositoryScanResultFilter struct {
 	Id           *string              `db:"id"`
 	RepositoryId *string              `db:"repository_id"`
-	Status       *grcenums.ScanStatus `db:"status"`
+	Status       *grenums.ScanStatus `db:"status"`
 }
