+++ grlib/db/gitrepository/v1/daos/git_repository_scan_result_with_detail.go
@@ -2,15 +2,15 @@ package grcgitrepositorydbdaos
 
 import (
 	"database/sql"
-	grcenums "github.com/kazekim/backend-engineer-challenge/grchallengeapi/enums"
 	"github.com/kazekim/backend-engineer-challenge/grlib/bejson"
+	"github.com/kazekim/backend-engineer-challenge/grlib/grenums"
 )
 
 type GitRepositoryScanResultWithDetail struct {
 	Id           *string              `db:"id"`
 	Name     string `db:"repository_name"`
 	Url      string `db:"repository_url"`
-	Status       grcenums.ScanStatus `db:"status"`
+	Status       grenums.ScanStatus `db:"status"`
 	Findings     bejson.JSON         `db:"findings"`
 	QueuedAt     sql.NullTime        `db:"queued_at"`
 	ScanningAt   sql.NullTime        `db:"scanning_at"`
