+++ grlib/db/gitrepository/v1/db_git_repository_update.go
@@ -1,8 +1,8 @@
 package grcgitrepositorydb
 
 import (
-	grcgitrepositorydbdaos "github.com/kazekim/backend-engineer-challenge/grchallengeapi/pkgs/db/gitrepository/v1/daos"
 	"github.com/kazekim/backend-engineer-challenge/grlib/besqlx"
+	grcgitrepositorydbdaos "github.com/kazekim/backend-engineer-challenge/grlib/db/gitrepository/v1/daos"
 	"github.com/kazekim/backend-engineer-challenge/grlib/grerrors"
 )
