+++ grlib/grenums/scan_status.go
@@ -1,4 +1,4 @@
-package grcenums
+package grenums
 
 type ScanStatus string
